% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csam-generics.R
\name{AIC.csam}
\alias{AIC.csam}
\title{Akaike's An Information Criterion}
\usage{
\method{AIC}{csam}(object, ..., k = 2)
}
\arguments{
\item{object}{A fitted object of class \code{"csam"} (must contain \code{Y}, \code{X},
\code{psi1}, and \code{psi2}).}

\item{...}{Currently unused.}

\item{k}{numeric, the penalty per parameter to be used; the default k = 2 is the classical AIC.}
}
\value{
A numeric value indicating the information criterion
}
\description{
S3 method for the base \code{stats::AIC} generic that computes AIC for a fitted \code{"csam"} object
If the penalty values \code{psi1} and \code{psi2}...
}
\examples{
\dontrun{
fit <- csam(Y, X, g = 2, d = 1, family = poisson(), trace = TRUE)
fit.aic <- AIC(fit)
}

}
