% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csam-generics.R
\name{BIC.csam}
\alias{BIC.csam}
\title{Bayesian Information Criterion}
\usage{
\method{BIC}{csam}(object, which.n.data = c("species", "sites", "all"), ...)
}
\arguments{
\item{object}{A fitted object of class \code{"csam"} (must contain \code{Y}, \code{X},
\code{psi1}, and \code{psi2}).}

\item{...}{Currently unused.}
}
\value{
A numeric value indicating the information criterion
}
\description{
S3 method for the base \code{stats::BIC} generic that computes BIC for a fitted \code{"csam"} object
If the penalty values \code{psi1} and \code{psi2}...
}
\examples{
\dontrun{
fit <- csam(Y, X, g = 2, d = 1, family = poisson(), trace = TRUE)
fit.aic <- BIC(fit)
}

}
