% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csam-helpers.R
\name{correct.uv}
\alias{correct.uv}
\title{Impose constraints on the factor scores and loadings}
\usage{
correct.uv(U, V)
}
\arguments{
\item{U}{A numeric matrix of factor scores of dimension \eqn{n \times d},
with rows corresponding to sites and columns to the dimension of the factor analytic term.}

\item{V}{A numeric matrix of factor loadings of dimension \eqn{s \times d},
with rows corresponding to species and columns to the dimension of the factor analytic term.}
}
\value{
A list of the constrained factor analytic terms with components:
\describe{
\item{u}{A numeric matrix of factor scores of dimension \eqn{n \times d}}
\item{v}{A numeric matrix of factor loadings of dimension \eqn{s \times d}}
}
}
\description{
Imposes constraints on the factor scores and loadings as in GMF.
}
\examples{
\dontrun{
set.seed(1)
n <- 50; s <- 10; p <- 2
X <- cbind(1, scale(matrix(rnorm(n * (p - 1)), n, p - 1)))
Y <- matrix(rpois(n * s, lambda = 5), n, s)
fit <- csam(Y, X, g = 2, d = 1, family = poisson(), trace = TRUE)
plot.csam(fit, param = "pll")
}

}
