% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csam-helpers.R
\name{estep_post_probs}
\alias{estep_post_probs}
\title{E-step posterior probabilities (internal)}
\usage{
estep_post_probs(
  Y,
  X,
  beta0,
  B,
  U,
  Lambda,
  phi,
  pi,
  family,
  trunc.interval = FALSE
)
}
\description{
E-step posterior probabilities (internal)
}
\keyword{internal}
