% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csam-helpers.R
\name{fitted_csam_internal}
\alias{fitted_csam_internal}
\title{Calculate the fitted values of a CSAM/SAM, \eqn{\mu_{ijk}}, and compute the posterior predictive mean (internal)}
\usage{
fitted_csam_internal(
  par.list,
  X,
  tau,
  family,
  type = c("response", "link"),
  archetype_specific = FALSE
)
}
\description{
Calculate the fitted values of a CSAM/SAM, \eqn{\mu_{ijk}}, and compute the posterior predictive mean (internal)
}
\keyword{internal}
