% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csam-helpers.R
\name{hessian_complete}
\alias{hessian_complete}
\title{Calculate the complete-data expected Hessian (internal)}
\usage{
hessian_complete(object)
}
\description{
Calculate the complete-data expected Hessian (internal)
}
\keyword{internal}
