% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csam-helpers.R
\name{mstep0_arch_pars}
\alias{mstep0_arch_pars}
\title{M-step for vanilla SAM: archetype parameters (internal)}
\usage{
mstep0_arch_pars(Y, X, beta0, B, phi, pi, tau, family, maxit = 1)
}
\description{
M-step for vanilla SAM: archetype parameters (internal)
}
\keyword{internal}
