% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csam-helpers.R
\name{mstep0_species_pars}
\alias{mstep0_species_pars}
\title{M-step for vanilla SAM: species parameters (internal)}
\usage{
mstep0_species_pars(Y, X, par.list, tau, family, maxit = 1)
}
\description{
M-step for vanilla SAM: species parameters (internal)
}
\keyword{internal}
