% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csam-helpers.R
\name{mstep_arch_pars}
\alias{mstep_arch_pars}
\title{M-step: archetype parameters (internal)}
\usage{
mstep_arch_pars(Y, X, beta0, B, U, Lambda, phi, pi, tau, family, maxit = 1)
}
\description{
M-step: archetype parameters (internal)
}
\keyword{internal}
