% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csam-helpers.R
\name{mstep_site_scores}
\alias{mstep_site_scores}
\title{M-step: site scores (internal)}
\usage{
mstep_site_scores(
  Y,
  X,
  B,
  beta0,
  Lambda,
  U,
  phi,
  pi,
  tau,
  family,
  psi1 = 0,
  maxit = 1
)
}
\description{
M-step: site scores (internal)
}
\keyword{internal}
