% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csam-helpers.R
\name{mstep_species_pars}
\alias{mstep_species_pars}
\title{M-step: species parameters (internal)}
\usage{
mstep_species_pars(
  Y,
  X,
  B,
  U,
  beta0,
  Lambda,
  phi,
  pi,
  tau,
  family,
  psi2 = 0,
  maxit = 1
)
}
\description{
M-step: species parameters (internal)
}
\keyword{internal}
