% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csam-helpers.R
\name{mzll}
\alias{mzll}
\title{Calculate the log-likelihood marginalised over the true archetype labels (internal)}
\usage{
mzll(Y, X, par.list, g = 3, d = 2, family = poisson(), psi1 = 0, psi2 = 0)
}
\description{
Calculate the log-likelihood marginalised over the true archetype labels (internal)
}
\keyword{internal}
