% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csam-helpers.R
\name{score_complete}
\alias{score_complete}
\title{Calculate the complete-data expected score function (internal)}
\usage{
score_complete(object)
}
\description{
Calculate the complete-data expected score function (internal)
}
\keyword{internal}
