% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csam-helpers.R
\name{se_csam}
\alias{se_csam}
\title{standard error estimator for csam fits}
\usage{
se_csam(object, U.random = TRUE, which.pars = NULL)
}
\arguments{
\item{object}{A fitted object of class \code{"csam"} (must contain \code{Y}, \code{X},
\code{psi1}, and \code{psi2})..}

\item{U.random}{Logical; indicating whether to treat site effects as fixed or random when using \code{method = "tmb"}.}

\item{which.pars}{Character; of parameter group name(s) to be returned. Must be a subset of the canonical parameters of a CSAM TMB fit.}
}
\value{
A two column data frame of estimates and corresponding standard errors
}
\description{
Computes standard errors for all model parameters in a fitted \code{"csam"} object using TMB's sdreport function.
}
\details{
The fitted \code{csam} object \strong{must} contain \code{Y}, \code{X}, and the penalty values
\code{psi1} and \code{psi2} (these are inferred from the object).
}
\examples{
\dontrun{
fit <- csam(Y, X, g = 2, d = 1, family = poisson(), trace = TRUE)
fit$Y <- Y; fit$X <- X
# csam now stores psi1 and psi2 internally
V_louis <- vcov(fit, method = "louis")
}

}
