% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csam-helpers.R
\name{tau_from_tmb_fit}
\alias{tau_from_tmb_fit}
\title{Calculate the posterior membership probabilities for a model fitted via TMB (internal)}
\usage{
tau_from_tmb_fit(Y, X, par.list, family)
}
\description{
Calculate the posterior membership probabilities for a model fitted via TMB (internal)
}
\keyword{internal}
